Extracted AkkaSubmissionsBridge
